creating code stream (but not using it yet)
the gif writer now builds the correct code stream out of the example
image data, but doesn't use it for writing the image data. turning the
stream into bytes is the next step.

now writing the file in-place instead of saving it to a buffer, because
there's no need to take up all that extra memory

memory is now being freed after I'm done with it, just for the sake of
good habits